requirements.txt
chromadb>=1.4.1
fastapi>=0.128.0
langchain-core>=0.3.31
langgraph>=0.2.44
openai>=2.15.0
opik>=1.9.87
pypdf>=6.6.0
python-dotenv>=1.2.1
requests>=2.32.5
sentence-transformers>=5.2.0
streamlit>=1.53.0
uvicorn>=0.40.0